basic go-linting // moved corecontext
